fix: Error checking
In several spots a value of -1 was being used to indicate an error
because it was outside the bounds of a valid triangle side length. This
doesn't work because -1 is also outside the bounds of an unsigned
integer. The program would have seen a side length of the maximum value
of an unsigned 32 bit integer as an invalid side length in this case. -1
has been replaced with 0 for error indication.